Lab 7

+++ app/Http/Controllers/ArticleController.php
@@ -6,12 +6,15 @@ use App\Models\Article;
 use App\Models\Comment;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+
+
     public function index()
     {
         $articles = Article::latest()->paginate(6);
@@ -31,6 +34,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', self::class);
         $request->validate([
             'date'=>'date',
             'name'=>'required|min:5|max:100',
@@ -68,6 +72,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $this->authorize('update', $article, self::class);
         $request->validate([
             'date'=>'date',
             'name'=>'required|min:5|max:100',
@@ -86,6 +91,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete',$article);
         if ($article->delete()) return redirect('/article')->with('status','Delete success');
         else return redirect()->route('article.show', ['article'=>$article->id])->with('status','Delete don`t success');
     }
